fix(cli): address PR review — move ABCs to niuu, fix manager bugs (NIU-399)
- Move ServicePlugin, Service, TUIPageSpec ABCs from cli/registry.py to
  niuu/ports/plugin.py per module boundary rules; cli re-exports them,
  volundr/tyr import from niuu.ports.plugin
- Remove dead code in _topological_sort (unused in_degree dict)
- Implement time-bounded health check using _health_check_timeout in
  _wait_healthy (was stored but never used)
- Fix test_settings_from_env to use monkeypatch.setenv instead of raw
  os.environ manipulation
